+++ parser/Cargo.toml
@@ -3,7 +3,10 @@ name = "parser"
 version = "0.1.0"
 edition = "2021"
 
+
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 lexer = { path = "../lexer" }
+#parsers = { path = "./parsers" }
