+++ parser/src/parsers/call_parser.rs
@@ -1,0 +1,5 @@
+use std::string::ParseError;
+
+trait CallParser<'a> {
+    fn call(&mut self) -> Result<Expr<'a>, ParseError>;
+}
